+++ src/components/pages/DAOcreator/index.tsx
@@ -15,7 +15,7 @@ import MembersStep from "./MembersStep";
 import SchemesStep from "./SchemesStep";
 import ReviewStep from "./ReviewStep";
 import DeployStep from "./DeployStep";
-import Support from "../../common/dao/Support";
+import Support from "../../common/Support";
 import { DAOForm } from "../../../lib/forms";
 import { FormState } from "formstate";
